refactor: add ConstShootScore Class
