+++ src/test/java/bowling/frame/ConstShootScore.java
@@ -1,0 +1,9 @@
+package bowling.frame;
+
+public final class ConstShootScore {
+
+    public static final ShootScore STRIKE = ShootScore.from(10);
+    public static final ShootScore FIVE_SCORE = ShootScore.from(5);
+    public static final ShootScore FOUR_SCORE = ShootScore.from(4);
+    public static final ShootScore ZERO_SCORE = ShootScore.from(0);
+}
