+++ src/test/java/bowling/frame/FramesTest.java
@@ -3,7 +3,7 @@ package bowling.frame;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import static bowling.frame.ShootScoreTest.FIVE_SCORE;
+import static bowling.frame.ConstShootScore.FIVE_SCORE;
 import static org.assertj.core.api.Assertions.*;
 
 class FramesTest {
